correction

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	float e;
 
 	/*size of variable types*/
-	printf("Size of a char: %2d byte(s)\n", sizeof(a));
-	printf("Size of an int: %2d byte(s)\n", sizeof(b));
-	printf("Size of a long int: %2d byte(s)\n", sizeof(c));
-	printf("Size of a long long int: %2d byte(s)\n" sizeof(d));
-	Printf("Size of a float: %2d byte(s)\n" sizeof(e));		
+	printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(a));
+	printf("Size of an int: %lu byte(s)\n",(unsigned long)sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n" (unsigned long)sizeof(d));
+	Printf("Size of a float: %lu byte(s)\n" (unsigned long)sizeof(e));		
 	return (0);
 
 }
